+++ R/jackknife.lm.R
@@ -7,9 +7,9 @@
 #' bias standard error, standard error and confidence intervals.
 #'
 #' @param formula Simple or multiple  linear regression formula with dependent and independent variables
-#' @param d Number of observations to be deleted from data to make jackknife samples
-#' @param data Data frame with dependent and independent independent variables
-#'    specified in the formula
+#' @param d Number of observations to be deleted from data to make jackknife samples. The default is 1 (for delete-1 jackknife).
+#' @param data Data frame with dependent and independent independent variables specified in the formula
+#' @param conf Confidence level, a positive number < 1. The default is 0.95.
 #' @return A list containing a summary data frame of jackknife estimates
 #'    with bias, standard error. t-statistics, and confidence intervals,
 #'    linear regression model of original data and a data frame with
@@ -25,26 +25,28 @@
 #' *Statistics & Probability Letters*, *6*(5), 341-347.
 #' \doi{10.1016/0167-7152(88)90011-9}
 #' @seealso [lm()] which is used for linear regression.
-#' @importFrom stats coefficients cor lm qnorm
+#' @importFrom stats coefficients lm qnorm
 #' @importFrom utils combn
 #' @export
 #' @examples
+#' ## library(jackknifeR)
 #' j.lm <- jackknife.lm(dist~speed, d = 2, data = cars)
 #' j.lm$jackknife.summary
 #' summary(j.lm$lm_mod)$coefficients
 #'
-jackknife.lm <- function(formula, d,  data){
+jackknife.lm <- function(formula, d = 1,  data, conf = 0.95){
 
   n <- nrow(data)
-
-  if((n*ncol(data))^d > 2e+08){message("This may take more time. Please wait...")}
+  if(is.numeric(conf)==FALSE||conf>1||conf<0) stop("Error: confidence level must be a numerical value between 0 and 1, e.g. 0.95")
+  if((n*ncol(data))^d > 9e+07) stop("The number of jackknife sub-samples will be huge")
+  if((n*ncol(data))^d > 1e+04){message("This may take more time. Please wait...")}
 
   cmb <- combn(n, d) # Row indexes to be eliminated for jackknife
   N <- ncol(cmb)     # Total number of jackknife samples
 
   lm_mod <- lm(formula, data = data) # Linear regression estimate
 
-  # A numeric vector to collect the jackknife estimates
+  # A data frame to collect the jackknife estimates
   jk <- as.data.frame(matrix(nrow = N, ncol = length(coefficients(lm_mod))))
   colnames(jk) <- names(coefficients(lm_mod))
 
@@ -59,8 +61,8 @@ jackknife.lm <- function(formula, d,  data){
   bias <- (n-d) * (theta_dot_hat-theta_hat) # Bias
   est <- theta_hat-bias
   jack_se <- sqrt((n-d)/d  *  rowMeans(apply(jk, 1, function(x) (x-theta_hat)^2))) # Jackknife standard error
-  jack_ci_lower <- est-(qnorm(0.975)*jack_se)
-  jack_ci_upper <- est+(qnorm(0.975)*jack_se)
+  jack_ci_lower <- est-(qnorm(0.5+(conf/2))*jack_se)
+  jack_ci_upper <- est+(qnorm(0.5+(conf/2))*jack_se)
 
   jackknife.summary <- data.frame(Estimate = est,
                                   bias = bias,
@@ -69,8 +71,16 @@ jackknife.lm <- function(formula, d,  data){
                                   ci.lower = jack_ci_lower,
                                   ci.upper = jack_ci_upper)
 
-  return(list(jackknife.summary = jackknife.summary,
-              lm_mod = lm_mod,
-              Jackknife.samples.coeff = jk))
+  jk.r <- list(jackknife.summary = jackknife.summary,
+               d = d,
+               conf.level = conf,
+               stat = formula,
+               n.jack = N,
+               original.estimate = theta_hat,
+               lm_mod = lm_mod,
+               Jackknife.samples.est = jk)
+  class(jk.r) <- "jk"
+
+  return(jk.r)
 }