+++ NEWS.md
@@ -1,7 +1,23 @@
+# jackknifeR 1.0.0
+
+* Version release
+
+# jackknifeR 0.2.0
+
+* Added summary.jk function
+* Added print.jk function
+* Added print.summary.jk function 
+
+# jackknifeR 0.1.2
+
+* Added specification of confidence level in the functions
+* Restricted the maximum number of sub-samples
+* Set default value for d 
+* Added a function for jackknifing in custom estimation function
+
 # jackknifeR 0.1.1
 
 * Added a `NEWS.md` file to track changes to the package.
 * Updated DOI formats
 * Added reference sources in DESCRIPTION
 * Checked for spelling errors
-